[STCC-4] Add support for Sensirion STCC-4 CO2 sensor
  - Merge SCD4x and STCC-4 driver
  - On boot, run reconditioning routine for STCC-4
  - Supports deep idle sleep

+++ src/driver_TIME.c
@@ -89,7 +89,7 @@ bool timerDelaySleepAsync_us(const uint32_t t_us, void (*cb)()) {
 }
 
 static bool timerDelaySleepLP(const uint16_t t_ms) {
-  uint32_t cc = ((t_ms * 1024u) / 1000u) - 1u;
+  uint32_t cc = (((uint32_t)t_ms * 1024u) / 1000u) - 1u;
   tdLPMatch   = false;
 
   TIMER_LP->COUNT16.CC[0].reg = cc;
